+++ spec/models/spree/gateway/adyen_hpp_spec.rb
@@ -4,11 +4,9 @@ describe Spree::Gateway::AdyenHPP do
   let(:hpp_source) { create :hpp_source, psp_reference: "9999" }
   let(:gateway) { described_class.new }
 
-  include_context "mock adyen api", success: true
-
   shared_examples "delayed gateway action" do
     context "when the action succeeds" do
-      include_context "mock adyen api", success: true
+      include_context "mock adyen client", success: true
 
       it { is_expected.to be_a ::ActiveMerchant::Billing::Response }
 
@@ -19,7 +17,7 @@ describe Spree::Gateway::AdyenHPP do
 
     context "when the action fails" do
       include_context(
-        "mock adyen api",
+        "mock adyen client",
         success: false,
         fault_message: "Should fail")
 
@@ -38,6 +36,24 @@ describe Spree::Gateway::AdyenHPP do
   describe ".credit" do
     subject { gateway.credit(2000, "9999", currency: "EUR") }
     include_examples "delayed gateway action"
+
+    include_context "mock adyen client", success: true
+
+    context "when additional data is provided" do
+      subject { gateway.credit(2000, "9999", currency: "EUR", additional_data: "TEST") }
+
+      it "includes them in the request" do
+        expect(client).to receive(:refund_payment).with(hash_including(:additional_data))
+        subject
+      end
+    end
+
+    context "when additional data is not provided" do
+      it "is not included in the request" do
+        expect(client).to_not receive(:refund_payment).with(hash_including(:additional_data))
+        subject
+      end
+    end
   end
 
   describe ".cancel" do